Move cross-platform integration-test scaffolding to shared-testkit
