+++ shared-testkit/src/main/scala/kinesis4cats/SSL.scala
