+++ shared-testkit/src/main/scala/kinesis4cats/logging/ConsoleLogger.scala
