+++ shared-testkit/src/main/scala/kinesis4cats/producer/ProducerSpec.scala
@@ -16,7 +16,6 @@
 
 package kinesis4cats
 package producer
-package fs2
 
 import scala.concurrent.duration._
 
@@ -32,16 +31,16 @@ import kinesis4cats.compat.retry.RetryPolicies._
 import kinesis4cats.compat.retry._
 import kinesis4cats.syntax.scalacheck._
 
-private[kinesis4cats] abstract class FS2ProducerSpec[PutReq, PutRes, A]
-    extends munit.CatsEffectSuite {
+private[kinesis4cats] abstract class ProducerSpec[PutReq, PutRes, A]
+    extends kinesis4cats.testkit.IntegrationSuite {
 
-  def producerResource: Resource[IO, FS2Producer[IO, PutReq, PutRes]]
+  def producerResource: Resource[IO, Producer[IO, PutReq, PutRes]]
   def streamName: String
   def aAsBytes(a: A): Array[Byte]
 
   def fixture(
       appName: String
-  ): SyncIO[FunFixture[FS2ProducerSpec.Resources[IO, PutReq, PutRes, A]]]
+  ): SyncIO[FunFixture[ProducerSpec.Resources[IO, PutReq, PutRes, A]]]
 
   override def munitIOTimeout: Duration = 5.minutes
 
@@ -60,8 +59,7 @@ private[kinesis4cats] abstract class FS2ProducerSpec[PutReq, PutRes, A]
           )
         )
       )
-      x <- records.traverse(resources.producer.put)
-      _ <- x.sequence_
+      _ <- resources.producer.put(records)
       retryPolicy = limitRetries[IO](30).join(constantDelay(1.second))
       size <- retryingOnFailures(
         retryPolicy,
@@ -80,11 +78,11 @@ private[kinesis4cats] abstract class FS2ProducerSpec[PutReq, PutRes, A]
   }
 }
 
-object FS2ProducerSpec {
+object ProducerSpec {
 
   final case class Resources[F[_], PutReq, PutRes, A](
       resultsQueue: Queue[F, A],
-      producer: FS2Producer[F, PutReq, PutRes]
+      producer: Producer[F, PutReq, PutRes]
   )
 
 }
