+++ shared-testkit/src/main/scala/kinesis4cats/syntax/scalacheck.scala
